[CHORE/#133] sideeffect 위치 이동
