--- app/src/main/java/com/cherrish/android/presentation/calendar/CalendarSideEffect.kt
@@ -1,7 +1,0 @@
-package com.cherrish.android.presentation.calendar
-
-import java.time.LocalDate
-
-sealed interface CalendarSideEffect {
-    data class NavigateToProcedure(val startDate: LocalDate) : CalendarSideEffect
-}
