+++ app/src/main/java/com/cherrish/android/presentation/calendar/CalendarUiState.kt
@@ -95,3 +95,7 @@ data class CalendarUiState(
         )
     }
 }
+
+sealed interface CalendarSideEffect {
+    data class NavigateToProcedure(val startDate: LocalDate) : CalendarSideEffect
+}
