--- app/src/main/java/com/cherrish/android/presentation/challenge/mission/ChallengeMissionSideEffect.kt
@@ -1,6 +1,0 @@
-package com.cherrish.android.presentation.challenge.mission
-
-sealed interface ChallengeMissionSideEffect {
-    data object NavigateToChallengeMissionProgress : ChallengeMissionSideEffect
-    data object NavigateToChallengeRoutine : ChallengeMissionSideEffect
-}
