+++ app/src/main/java/com/cherrish/android/presentation/challenge/mission/ChallengeMissionUiState.kt
@@ -49,3 +49,9 @@ data class ChallengeMissionUiState(
         )
     }
 }
+
+sealed interface ChallengeMissionSideEffect {
+    data object NavigateToChallengeMissionProgress : ChallengeMissionSideEffect
+    data object NavigateToChallengeRoutine : ChallengeMissionSideEffect
+}
+
